requirements.txt
json
pandas
matplotlib
pillow
streamlit_authenticator